# backend/docs/swagger.yaml
basePath: /
definitions:
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
...
  pgtype.Timestamp:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        description: Time zone will be ignored when encoding to PostgreSQL.
        type: string
      valid:
        type: boolean
    type: object
  server.CreateDiseaseRequest:
    properties:
      disease_code:
        type: string
      disease_description:
        type: string
      disease_name:
        type: string
      disease_treatment:
        description: Accept raw JSON
        items:
          type: integer
        type: array
    type: object
...
  server.DiseaseResponse:
    properties:
      created_at:
        allOf:
        - $ref: '#/definitions/pgtype.Timestamp'
        description: Or format as string
      disease_code:
        type: string
      disease_description:
        type: string
      disease_id:
        type: integer
      disease_name:
        type: string
      disease_treatment:
        description: Send raw JSON back
        items:
          type: integer
        type: array
      updated_at:
        allOf:
        - $ref: '#/definitions/pgtype.Timestamp'
        description: Or format as string
    type: object
  server.HTTPError:
    properties:
      error:
        description: Error message
        example: Resource not found
        type: string
    type: object
...
  server.UpdateDiseaseRequest:
    properties:
      disease_code:
        type: string
      disease_description:
        type: string
      disease_name:
        type: string
      disease_treatment:
        description: Accept raw JSON
        items:
          type: integer
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Service for managing patients, symptoms, and diseases.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Patient API Service
  version: "1.0"
paths:
...
  /diseases:
    get:
      consumes:
      - application/json
      description: Get a list of all diseases
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of diseases
          schema:
            items:
              $ref: '#/definitions/server.DiseaseResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: List diseases
      tags:
      - Diseases
    post:
      consumes:
      - application/json
      description: Add a new disease record
      parameters:
      - description: Disease data to create
        in: body
        name: disease
        required: true
        schema:
          $ref: '#/definitions/server.CreateDiseaseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Disease created successfully
          schema:
            $ref: '#/definitions/server.DiseaseResponse'
        "400":
          description: Invalid request payload or validation error
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Create a new disease
      tags:
      - Diseases
  /diseases/{diseaseID}:
    delete:
      consumes:
      - application/json
      description: Delete a disease record by ID
      parameters:
      - description: Disease ID
        format: int32
        in: path
        name: diseaseID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content (Successful deletion)
          schema:
            type: string
        "400":
          description: Invalid Disease ID format
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Delete disease
      tags:
      - Diseases
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific disease by its ID
      parameters:
      - description: Disease ID
        format: int32
        in: path
        name: diseaseID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved disease
          schema:
            $ref: '#/definitions/server.DiseaseResponse'
        "400":
          description: Invalid Disease ID format
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Disease not found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Get disease by ID
      tags:
      - Diseases
    put:
      consumes:
      - application/json
      description: Update details for an existing disease
      parameters:
      - description: Disease ID
        format: int32
        in: path
        name: diseaseID
        required: true
        type: integer
      - description: Disease data to update
        in: body
        name: disease
        required: true
        schema:
          $ref: '#/definitions/server.UpdateDiseaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Disease updated successfully
          schema:
            $ref: '#/definitions/server.DiseaseResponse'
        "400":
          description: Invalid request payload or validation error
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Disease not found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Update disease details
      tags:
      - Diseases
...
schemes:
- http
- https
swagger: "2.0"